Ajout message de confirmation pour les suppressions de produits

+++ TP3/assets/scripts/controllers/ShoppingCartController.js
@@ -53,20 +53,24 @@ ShoppingCartController.prototype = {
   }, 
 
   removeProduct : function(productId) {
-    this._model.removeProduct(productId);
-    this._headerController.setNumberOfProducts(this._model.getNumberOfProducts());
-    localStorage.setItem("shoppingCart",JSON.stringify(this._model.getShoppingCart()));
-    if(this._view && this._model.getNumberOfProducts() === 0) {
-      this._view.showMessageError(this._messages.noProductInCart);
+    if(this._view.showModalConfirmMessage(this._messages.askToRemoveProduct)) {
+      this._model.removeProduct(productId);
+      this._headerController.setNumberOfProducts(this._model.getNumberOfProducts());
+      localStorage.setItem("shoppingCart",JSON.stringify(this._model.getShoppingCart()));
+      if(this._view && this._model.getNumberOfProducts() === 0) {
+        this._view.showMessageError(this._messages.noProductInCart);
+      } 
     } 
   }, 
 
   removeAllProducts : function() {
-    this._model.removeAllProducts();
-    this._headerController.setNumberOfProducts(0);
-    localStorage.removeItem("shoppingCart",JSON.stringify());
-    if(this._view) {
-      this._view.showMessageError(this._messages.noProductInCart);
+    if(this._view.showModalConfirmMessage(this._messages.askToEmptyShoppingCart)) {
+      this._model.removeAllProducts();
+      this._headerController.setNumberOfProducts(0);
+      localStorage.removeItem("shoppingCart",JSON.stringify());
+      if(this._view) {
+        this._view.showMessageError(this._messages.noProductInCart);
+      }
     } 
   },
 
@@ -74,9 +78,6 @@ ShoppingCartController.prototype = {
     this._model.changeProductQuantity(productId, deltaQuantity);
     this._headerController.setNumberOfProducts(this._model.getNumberOfProducts());
     localStorage.setItem("shoppingCart",JSON.stringify(this._model.getShoppingCart()));
-    // if(this._view && this._model.getNumberOfProducts() === 0) {
-    //   this._view.showMessageError(this._messages.noProductInCart);
-    // } 
   }
 
 };
